feat(background): added canvas object to site background. (#78)
* feat(background): added canvas object to site background.

* feat(background): updated background animation and now fully functions.

Now it is possible to simply add more items to the background and the
rendering happens as expected. The background also can resize and
recreates the space background.

* fix(shooting-stars): fixed shooting star chances and load.

* feat(planets): added planets and refactored some code.

* feat(background): added labels to planets with random names.

Also temporarily removed the moons as it is a little trickier
retrieving the current position of the planets, but I do have
a fix, it lies in the planets draw function. I also made the
sun bigger and slowed down the rotation of the objects.

* test(background): added tests to cover regression of space background animation.

Renamed randomNameGenerator to getRandomName.